REMOVE WP BOOTSTRAP STARTER CHILD THEME FILES
The site is currently based on the wp-bootstrap-starter theme,
customised via a wp-bootstrap-starter-child theme. As part of the
redesign, and upgrade of the WordPress version used, I'm changing to the
twentytwentythree theme customised via a twentytwentythree-child theme.
So, we'll start by removing the wp-bootstrap-starter-child theme files.
Where the new twentytwentythree-child theme as files of the same name
in its hierachy we won't track them as changes to the
wp-bootstrap-starter-child files as they're not really different
versions of the same thing.