INITIAL CHILD THEME CUSTOMISATIONS OF HEADER AND FOOTER TEMPLATE PARTS
Note that the customised header template part makes reference to the
navigation menu that appears in the header using its corresponding post
ID. This identifier won't necessarily be the same across different
environments. I have to think about how to manage this within a version
controlled file that will be used to deploy to multiple environments.