INITIAL CHILD THEME FRONT PAGE TEMPLATE
Configure the page to display the content of the front page and then a
grid query loop of posts. I've just used the starting query loop for now
but the posts shown here will be upcoming events, which is likely to
require a custom post type. So, that change is coming.