INITIAL CHILD THEME PAGE TEMPATES FOR SITE SECTIONS
These are the initial child theme page templates for the "Visit the
Viaduct" and "Support the FoBV" site sections. They only vary from the
standard Page template in having a navigation section immediately below
the page header. This is for navigation between pages within the
section.

Again these make references to the associated menu posts by ID, which is
problematic in a version controlled file that I want to use to deploy to
different environments, where that ID will vary. I'll have to work out
an approach to this problem later.