THE FOBV EVENT PLUGIN AS SCAFFOLDED BY THE WORDPRESS CREATE BLOCK TOOL
I am committing the inital state of this plugin as generated by the
WordPress Create Block tool with the exception of its build folder so
that my changes to it are tracked. I don't see why I should want to
track the build folder, since it contains only generated content, though
the .gitignore in the scaffold only excludes build/Release and not build
folder files outside of the Release subfolder.