MOVE NODE DEPENDENCIES INTO THEME FOLDER
I was using a single set of node dependencies for the overall wordpress
repository. However, it becomes clear that we're going to need these at
various points within the hierarchy of the wordpress repository; for
example, for the theme, for the fobv-event plugin, etc. The one that was
at the top-level in the repository is in reality specific to the theme.