DELETE UNNECESSARY FILES FROM THE FOBV SITE PLUGIN
I used the WordPress Create Block tool to scaffold this plugin. However,
as it turns out I have not needed to register a custom block because I
have instead been able to achieve what I wanted by registering a block
variation on the query loop block. This means that there are a number of
files that were generated by the Create Block tool that I now don't need
and so I am deleting them accordingly.