MERGE START DATE AND END DATE COMPONENTS OF THE FOBV EVENT PLUGIN
I need to share at least one Javascript function across both the start
date and end date so the quickest and easiest way to do this is just to
combine those two components into a single "Event Dates" component. I'm
committing that new component based on just a rename of the existing
start date component so that we can then track changes from that point.