CATER FOR DUPLICATE THEME IMAGES IN THE MEDIA LIBRARY IN THE INIT SCRIPT
We may end up with duplicate theme images in the media library, for
whatever reason. When this happens, those images added after the first
will be differentiated with a suffix consisting of a hyphen followed by
an underscore.

The init.sh WP-CLI script deletes theme images from the media library
before re-adding them. This change enables the delete step to identify
and also remove duplicates.