RENAME SUPPORT THE FOBV SECTION PAGE TEMPLATE
This section has been renamed from "Support the FoBV" to "Support our
Charity" so the template file is renamed accordingly. There may be
further change to follow for the same reason but right now I'm just
committing the renaming action so that further change can be tracked
from that point.