REFLECT SUPPORT THE FOBV SECTION NAME CHANGE IN INIT SCRIPT & THEME JSON
Reference the description for the previous commit, these are knock on
consequences of changing the "Support the FoBV" page and section to
"Support our Charity" in both the WP-CLI init.sh script and the theme's
theme.json file. It turns out the the template file that we renamed in
the previous commit did not require further change as a consequence.