USE CUSTOM MENU TITLE FOR THE HOME PAGE WHEN CREATING LEGACY MENUS
Since it seems to be quite difficult to version control the new, block
based menus, we're version controlling classical menus that can then be
imported as block menus. By default the menu item titles are the same as
their corresponding page titles but the home page is different, the
page title is "Bennerley Viaduct: The Iron Giant" but we use the title
"Home" for the corresponding menu title. Building that into the creation
of the legacy menus means that we don't then have to manually make that
change after we have imported them.