REVERT HEADER LOGO IMAGE TO A TRANSPARENT BACKGROUND
I had changed the generation of this image so that a background layer
was added that was filled with the same colour as that used to fill the
header section that the image is placed within. For some reason, which
I've not got to the bottom of, there was a very minor difference between
the background colour of the image and the background colour of the
header section discernable to the eye. This despite investigation
confirming that both used the same colour code. Rather than put further
work into cracking why this is, I've reverted to a transparent
background to the image.