GREATER FLEXIBILITY IN THE FUNCTION TO OUTPUT A PAYMENT REFERENCE
This function enabled us to generate and output a payment reference as
either a hidden input field or as plain text on the FIRST page of a
process flow only. These changes allow it to also output the payment
reference associated with the transaction that is initiated on the first
page of a process flow on the subsequent pages in that process flow. The
function has had extensive comments added to make its function clearer.