VERIFY USER RECAPTCHA RESPONSES
I had decided at go-live not to take any action initially based on
reCAPTCHA responses but instead to monitor them before deciding what to
do. In doing this, I wasn't verifying the reCAPTCHA user responses,
which the reCAPTCHA admin console highlights as an issue, because it
knows when it's issued a token and not seen a subsequent verfiication.

So, I've now added that verification but still as yet without any action
based on the result. I've also hooked notification of the verification
result to me via email, which will help to understand what's happening
before implementing the correct course of action for suspicious
requests.