CHANGES TO RECAPTCHA PROCESSING WITHIN SITE FORMS THAT USE IT
Two changes as follows:

 - Put the Javascript to handle reCAPTCHA within the form's Javascript
   file rather than as a custom HTML block associated with the form.

 - Change from the "Automatically bind the challenge to a button"
   approach to the "Programatically invoke the challenge" approach.

The two approaches in the second of those two bullet points are
described here:
https://developers.google.com/recaptcha/docs/v3

I changed the approach because I SEEMED to be having issues with the
"Automatically bind the challenge to a button" approach in respect of:

 - Having two forms that are both protected by reCAPTCHA on the same
   page.

 - Having reCAPTCHA work harmoniously alongside jQuery Validation.

In hindsight, I'm not 100% sure that either of those were fixed by
chaging the approach. However, both issues are fixed now and I find that
I prefer the "Progrmatically invoke the challenge" approach because what
is happening is more transparent, to me at least.