CLEAR PAYMENT REFERENCE IN BANK TRANSFER PAYMENT INSTRUCTIONS
In converting the cheque and bank transfer payment instructions to
reusable blocks, I had forgotten the need clear the payment reference
from the session after each form so that a new reference is allocated to
the next form instance. Note that re-adding this here relies on the fact
that this is the last use of the payment reference in each of the
"Donate" and "Join Us" forms. A more robust approach would be to be able
to include the fobv-payment-reference simply to clear the reference from
the session without outputting anything - see issue #37.